+++ ChangeLog.md
@@ -8,13 +8,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- split `GraphDocumentReadOptions` from `DocumentReadOptions` (breaking change)
 - added `ArangoCollection#getResponsibleShard(Object)`
 - added support for Analyzers
 - added support for Stream Transactions
 - added support for named indices
 - added support for TTL indices
 - added minReplicationAttribute for collections and graphs
-- added batched thread support in ArangoCollection.importDocuments (by @rkhaja)
 
 ## [5.0.7] - 2019-07-19